+++ fixtures/regressions/logging-callback-interface/tests/test_generated_bindings.rs
@@ -1,0 +1,1 @@
+uniffi_macros::build_foreign_language_testcases!("tests/bindings/test.py",);